Refactored the table in the tracker page into a seperate components Created new ProjectTabel component to hold the project table
